pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aeeprotocol"
version = "0.5.0"
description = "Quantum-Resistant Data Sovereignty Protocol with Kyber-768"
readme = "README.md"
authors = [{name = "Franco Luciano Carricondo", email = "tu-email@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "liboqs-python>=0.9.0",
    "numpy>=1.21.0",
    "cryptography>=41.0.0"
]

[project.optional-dependencies]
api = ["flask>=3.0.0", "flask-cors>=4.0.0"]
dev = ["pytest>=7.0.0", "black>=23.0.0", "flake8>=6.0.0"]
demo = ["fastapi>=0.104.0", "uvicorn>=0.24.0"]

[project.urls]
Homepage = "https://github.com/ProtocoloAEE/aee-protocol"

[tool.setuptools.packages.find]
where = ["src"]
include = ["aeeprotocol*"]

[tool.setuptools.package-data]
"aeeprotocol" = ["py.typed"]